refactor computer to check mem safe regions
