+++ memsafe-checker/src/common.rs
@@ -1,11 +1,11 @@
 use std::fmt;
 use std::io::Write;
+use std::str::FromStr;
 
-#[derive(Debug, Clone, Copy)]
+#[derive(Debug, Clone, Copy, Eq, PartialEq)]
 pub enum RegionType {
     READ,
     WRITE,
-    READWRITE,
 }
 
 impl fmt::Display for RegionType {
@@ -13,7 +13,6 @@ impl fmt::Display for RegionType {
         match self {
             RegionType::READ => write!(f, "Read"),
             RegionType::WRITE => write!(f, "Write"),
-            RegionType::READWRITE => write!(f, "Read/Write"),
         }
     }
 }
@@ -21,9 +20,136 @@ impl fmt::Display for RegionType {
 // TODO: allow different types within one region at different offsets
 // for example for sha256 input is also output may be 1000 bits long but output is 256
 // so shouldn't write into the input buffer past 256 bits
+#[derive(Debug, Clone)]
 pub struct MemorySafeRegion {
     pub region_type: RegionType,
     pub register: String,
     pub start_offset: usize,
     pub end_offset: usize,
+}
+
+#[derive(Debug, Clone)]
+pub struct MemorySafetyError {
+    details: String,
+}
+
+impl MemorySafetyError {
+    pub fn new(msg: &str) -> MemorySafetyError {
+        MemorySafetyError {
+            details: msg.to_string(),
+        }
+    }
+}
+impl fmt::Display for MemorySafetyError {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        write!(f, "{}", self.details)
+    }
+}
+
+#[derive(Debug, Clone)]
+pub struct Instruction {
+    pub op: String,
+    pub r1: Option<String>,
+    pub r2: Option<String>,
+    pub r3: Option<String>,
+    pub r4: Option<String>,
+}
+
+impl Instruction {
+    pub fn new(text: String) -> Instruction {
+        Instruction {
+            op: text,
+            r1: None,
+            r2: None,
+            r3: None,
+            r4: None,
+        }
+    }
+}
+#[derive(Debug, Clone)]
+pub struct ParseInstructionError;
+
+impl FromStr for Instruction {
+    type Err = ParseInstructionError;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        //find if there's anything in brackets to allow fun addressing modes
+        let mut brac: String = Default::default();
+        if s.contains("[") {
+            let left = s.find('[');
+            let right = s.rfind(']');
+            let exclamation = s.rfind('!');
+            if left.is_some() && right.is_some() {
+                brac = s[left.unwrap()..right.unwrap()].to_string();
+            }
+            if exclamation.is_some() {
+                brac = brac + "!";
+            }
+        }
+
+        let v: Vec<&str> = s.split(|c| c == '\t' || c == ',' || c == ' ').collect();
+
+        let v0 = v[0].to_string();
+        let v1: Option<String>;
+        let v2: Option<String>;
+        let v3: Option<String>;
+        let v4: Option<String>;
+
+        if v.len() > 1 {
+            let val1 = v[1].to_string();
+            if val1.contains("[") {
+                v1 = Some(brac.clone());
+            } else if val1.contains("]") {
+                v1 = None;
+            } else {
+                v1 = Some(val1);
+            }
+        } else {
+            v1 = None;
+        }
+        if v.len() > 2 {
+            let val2 = v[2].to_string();
+            if val2.contains("[") {
+                v2 = Some(brac.clone());
+            } else if val2.contains("]") {
+                v2 = None;
+            } else {
+                v2 = Some(val2);
+            }
+        } else {
+            v2 = None;
+        }
+        if v.len() > 3 && !v[3].is_empty() {
+            let val3 = v[3].to_string();
+            if val3.contains("[") {
+                v3 = Some(brac.clone());
+            } else if val3.contains("]") {
+                v3 = None;
+            } else {
+                v3 = Some(val3);
+            }
+        } else {
+            v3 = None;
+        }
+        if v.len() > 4 && !v[4].is_empty() {
+            let val4 = v[4].to_string();
+            if val4.contains("[") {
+                v4 = Some(brac);
+            } else if val4.contains("]") {
+                v4 = None;
+            } else {
+                v4 = Some(val4);
+            }
+        } else {
+            v4 = None;
+        }
+
+        Ok(Instruction {
+            op: v0,
+            r1: v1,
+            r2: v2,
+            r3: v3,
+            r4: v4,
+        })
+    }
 }
